Move example invoice from InvoiceEditor to seeds
